+++ packages/react-router/__tests__/utils/getWindow.ts
@@ -1,0 +1,8 @@
+import { JSDOM } from "jsdom";
+
+export default function getWindow(initialUrl: string, isHash = false): Window {
+  // Need to use our own custom DOM in order to get a working history
+  const dom = new JSDOM(`<!DOCTYPE html>`, { url: "http://localhost/" });
+  dom.window.history.replaceState(null, "", (isHash ? "#" : "") + initialUrl);
+  return dom.window as unknown as Window;
+}