+++ packages/react-router/lib/router/router.ts
@@ -3647,8 +3647,28 @@ export function createStaticHandler(
             return res;
           },
           async (error, routeId) => {
-            if (isResponse(error)) {
+            // Redirects propagate verbatim
+            if (isRedirectResponse(error)) {
               return error;
+            }
+
+            // All other thrown responses during document/data request bubble
+            // to the boundary
+            if (isResponse(error)) {
+              try {
+                error = new ErrorResponseImpl(
+                  error.status,
+                  error.statusText,
+                  await parseResponseBody(error),
+                );
+              } catch (e) {
+                error = e;
+              }
+            }
+
+            if (isDataWithResponseInit(error)) {
+              // Convert thrown data() values to ErrorResponses for the UI
+              error = dataWithResponseInitToErrorResponse(error);
             }
 
             if (renderedStaticContext) {
@@ -3850,8 +3870,9 @@ export function createStaticHandler(
           return res;
         },
         (error) => {
-          if (isRouteErrorResponse(error)) {
-            return Promise.resolve(errorResponseToResponse(error));
+          if (isDataWithResponseInit(error)) {
+            // Convert thrown data() values to Responses for resource routes
+            return Promise.resolve(dataWithResponseInitToResponse(error));
           }
           if (isResponse(error)) {
             return Promise.resolve(error);
@@ -4314,10 +4335,19 @@ export function createStaticHandler(
             basename,
           );
         }
-        if (isResponse(result.result) && isRouteRequest) {
-          // For SSR single-route requests, we want to hand Responses back
-          // directly without unwrapping
-          throw result;
+
+        // For SSR single-route requests, we want to hand Responses back
+        // directly, as well as upgrade data() calls to Response instances
+        // (this allows utilities using data() and be shared between normal and
+        // resource routes).
+        if (isRouteRequest) {
+          if (isResponse(result.result)) {
+            throw result;
+          } else if (isDataWithResponseInit(result.result)) {
+            // Upgrade `data()` to `Response` so utilities using `data()` can be
+            // shared between resource and non-resource routes
+            throw dataWithResponseInitToResponse(result.result);
+          }
         }
 
         dataResults[match.route.id] =
@@ -5399,13 +5429,7 @@ function runServerMiddlewarePipeline(
     isResponse,
     async (error, routeId, nextResult) => {
       // Convert thrown data() values to ErrorResponses
-      let response = await errorHandler(
-        isDataWithResponseInit(error)
-          ? dataWithResponseInitToErrorResponse(error)
-          : error,
-        routeId,
-        nextResult,
-      );
+      let response = await errorHandler(error, routeId, nextResult);
       return response;
     },
   );
@@ -5930,6 +5954,18 @@ async function callLoaderOrAction({
   return result;
 }
 
+async function parseResponseBody(response: Response) {
+  let contentType = response.headers.get("Content-Type");
+
+  // Check between word boundaries instead of startsWith() due to the last
+  // paragraph of https://httpwg.org/specs/rfc9110.html#field.content-type
+  if (contentType && /\bapplication\/json\b/.test(contentType)) {
+    return response.body == null ? null : response.json();
+  }
+
+  return response.text();
+}
+
 async function convertDataStrategyResultToDataResult(
   dataStrategyResult: DataStrategyResult,
 ): Promise<DataResult> {
@@ -5939,18 +5975,7 @@ async function convertDataStrategyResultToDataResult(
     let data: any;
 
     try {
-      let contentType = result.headers.get("Content-Type");
-      // Check between word boundaries instead of startsWith() due to the last
-      // paragraph of https://httpwg.org/specs/rfc9110.html#field.content-type
-      if (contentType && /\bapplication\/json\b/.test(contentType)) {
-        if (result.body == null) {
-          data = null;
-        } else {
-          data = await result.json();
-        }
-      } else {
-        data = await result.text();
-      }
+      data = await parseResponseBody(result);
     } catch (e) {
       return { type: ResultType.error, error: e };
     }
@@ -6482,10 +6507,7 @@ function isHashChangeOnly(a: Location, b: Location): boolean {
 function dataWithResponseInitToResponse<D>(
   data: DataWithResponseInit<D>,
 ): Response {
-  return new Response(
-    typeof data.data === "string" ? data.data : JSON.stringify(data.data),
-    data.init || undefined,
-  );
+  return Response.json(data.data, data.init ?? undefined);
 }
 
 function dataWithResponseInitToErrorResponse<D>(
@@ -6495,16 +6517,6 @@ function dataWithResponseInitToErrorResponse<D>(
     data.init?.status ?? 500,
     data.init?.statusText ?? "Internal Server Error",
     data.data,
-  );
-}
-
-function errorResponseToResponse(error: ErrorResponse): Response {
-  return new Response(
-    typeof error.data === "string" ? error.data : JSON.stringify(error.data),
-    {
-      status: error.status,
-      statusText: error.statusText,
-    },
   );
 }